repo: add flag to remember if singleton is destroyed
pytorch for example would free hipFFT plan handles at process exit.
Unfortunately, we have a potential ordering problem because librocfft.so is
also unloaded at process exit.  If the unload happens first, then the Repo
singleton is destroyed, and the subsequent freeing of hipFFT plan handles
will crash.

Instead, keep track of when the singleton is destroyed - this should only
happen when the library is unloaded, which means the repo will clean up all
of the stored plans anyway, and the freeing can then become a no-op.