add html_report.py to generate performance graphs using plotly
* scripts: add script to generate html report using plotly

This is more interactive and gives better visibility into the actual
problem sizes being measured.  It currently only draws one line per
directory (as opposed to one line per radix), plus a line for speedup.

The idea is that you would give a set of problem sizes to test that aren't
necessarily just one radix.

* scripts: factor radix size generator out to separate function

To make it easier to swap another generator in.

* scripts: add -F option to get problem sizes from file

Currently, if you specify this you will get an html perf report (not pdf,
docx).